Continue fixing merge

+++ autodiscovery/pkg/peerHeartbeatState.go
@@ -7,13 +7,13 @@ import (
 //PeerHeartbeatState describes the living state of a peer
 type PeerHeartbeatState interface {
 	GenerationTime() time.Time
-	ElapsedHeartbeats() uint64
+	ElapsedHeartbeats() int64
 	MoreRecentThan(PeerHeartbeatState) bool
 }
 
 type heartbeatState struct {
 	generationTime    time.Time
-	elapsedHeartbeats uint64
+	elapsedHeartbeats int64
 }
 
 //GenerationTime returns the peer's generation time
@@ -22,7 +22,7 @@ func (hb heartbeatState) GenerationTime() time.Time {
 }
 
 //ElapsedHeartbeats returns the peer's elapsed living seconds from the latest refresh
-func (hb heartbeatState) ElapsedHeartbeats() uint64 {
+func (hb heartbeatState) ElapsedHeartbeats() int64 {
 	if hb.elapsedHeartbeats == 0 {
 		hb.refreshElapsedHeartbeats()
 	}
@@ -30,7 +30,7 @@ func (hb heartbeatState) ElapsedHeartbeats() uint64 {
 }
 
 func (hb *heartbeatState) refreshElapsedHeartbeats() {
-	hb.elapsedHeartbeats = uint64(time.Now().Unix() - hb.generationTime.Unix())
+	hb.elapsedHeartbeats = time.Now().Unix() - hb.generationTime.Unix()
 }
 
 //RecentThan check if the current heartbeat state is more recent than the another heartbeat state
@@ -53,7 +53,7 @@ func (hb heartbeatState) MoreRecentThan(hbS PeerHeartbeatState) bool {
 }
 
 //NewPeerHeartbeatState creates a new peer's heartbeat state
-func NewPeerHeartbeatState(genTime time.Time, elapsedHb uint64) PeerHeartbeatState {
+func NewPeerHeartbeatState(genTime time.Time, elapsedHb int64) PeerHeartbeatState {
 	return heartbeatState{
 		generationTime:    genTime,
 		elapsedHeartbeats: elapsedHb,
